+++ app/components/select-month/component.js
@@ -1,0 +1,5 @@
+import Ember from 'ember';
+
+export default Ember.Component.extend({
+  months: ['01', '02', '03', '04', '05', '06', '07', '08', '09', '10', '11', '12']
+});
