+++ app/components/select-month/template.hbs
@@ -1,0 +1,1 @@
+{{view "select" content=months value=value class="form-control" name=name}}
