+++ app/components/select-plan/component.js
@@ -1,0 +1,5 @@
+import Ember from 'ember';
+
+export default Ember.Component.extend({
+  plans: ['development', 'pilot', 'platform', 'production']
+});
