+++ app/components/select-plan/template.hbs
@@ -1,0 +1,1 @@
+{{view "select" content=plans value=value class="form-control" prompt="Select a plan" name=name}}
