+++ app/components/select-year/component.js
@@ -1,0 +1,5 @@
+import Ember from 'ember';
+
+export default Ember.Component.extend({
+  years: ['2015', '2016', '2017', '2018', '2019', '2020']
+});
