+++ app/components/select-year/template.hbs
@@ -1,0 +1,1 @@
+{{view "select" content=years value=value class="form-control" name=name}}
