+++ tests/unit/components/select-year-test.js
@@ -1,0 +1,25 @@
+import {
+  moduleForComponent,
+  test
+} from 'ember-qunit';
+
+import Ember from 'ember';
+
+moduleForComponent('select-year', 'SelectYearComponent', {
+  needs: [],
+  setup: function() {
+    this.container.register('view:select', Ember.Select);
+  }
+});
+
+test('it renders', function() {
+  expect(2);
+
+  // creates the component instance
+  var component = this.subject({});
+  equal(component._state, 'preRender');
+
+  // appends the component to the page
+  this.append();
+  equal(component._state, 'inDOM');
+});
